requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==25.0.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
